visual: extract flags from vmenus.c to flags.c
